Added crud and transformer

+++ src/main/java/com/project/races/dto/user/RaceResponse.java
@@ -23,5 +23,5 @@ public class RaceResponse {
 
     private String dateOfStart;
 
-    private List<Integer> teamsID;
+    private List<Long> teamsID;
 }
